[Fixed] Command Line Parsing of Arguments
Fixes a minor oversight in the command line interface code that caused the command line arguments to not be parsed out before the CompilationState structure was determined.
Creates a templated "OptionState<T>" type (name TBD) to allow for storing the parsed options without introducing unneeded dependencies (for the moment).
Utilizes the new type to merge the command line parsing into a single function that works as expected