Implemented Assignment and Operator Codegen
Adds the necessary hook changes to enable codegen for binary/unary operations and assignment
Fixes issue with the test runner where speroc was being called with a deprecated flag
Fixes issue where the compiler queried a modified command flag, causing crashes
NOTE: There is currently an issue with assignments and variables. Using the included interpreter, the code performs as expected. However when running an executable created from the command line, a return of 1 is given