Hide Typing Phase In Compilation
Temporarily hides the typing phase from the compilation passes. The recent work is just a basic API search, to see what structures are needed and what approaches work for solving the problem. More work will be done when it comes time to implement the type inference algorithm, however that requires work on functions first.