Fix Issue With Recursion Tests (due to Ruby)
For some reason, $?.exitstatus is defined to only maintain the lower 8-bits of the exit status. This was causing the recursion test's return value to be misreported. Since it's explicitly mentioned in the ruby documentation, I decided to instead print an explicit warning if the expected value can not be represented